nvme/perf: Fix SPDK perf issue with emulated NVMe PCIe controller
Emulated NVMe controller does not support Get Features "Number of
Queues" command. The same information can be retrieved from completion
for Set Features "Number of Queues" command.

It is not clear from NVMe specification if Get Features "Number of
Queues" is mandatory. If yes, NVMe emulation should be fixed to
support it.

Signed-off-by: Evgeniy Kochetov <evgeniik@mellanox.com>