rdma: Allocate a context with a pointer to qpair when handling LAST_WQE_REACHED event
The following scenario might occur when nvmf_tgt is stopped:
1. nvmf_tgt receives SIGINT, changes state to NVMF_TGT_FINI_STOP_SUBSYSTEMS
2. In this state nvmf_tgt stops all subsystems and disconnects associated qpairs
3. In the case of RDMA qpair, its state will be changed to IBV_QPS_ERR.
Once qpair changes the state to IBV_QPS_ERR, RDMA device generates
LAST_WQE_REACHED event when there are no more WQE that can be sonsumed
from the SRQ by this qpair.
4. When all subsystems are stopped, some of qpair may still be alive since they
haven't received LAST_WQE_REACHED event yet.
5. nvmf_tgt stops all poll groups and forcefully destroyes any qpairs linked to them.
6. At this moment LAST_WQE_REACHED event might be generated and received in another thread.
Handler of this event sends a message with a pointer to qpair. The qpair itself may already
be destroyed.
7. Thread that owned qpair receives a message (LAST_WQE_REACHED) with a pointer to alredy destroyed qpair and
destroyes it for the second time when all pointer are invalid.

Introduce a new structure with a pointer to the qpair, this structure will be sent as a message context
to qpair owner thread; add a pointer to this structure to the qpair itself.
Zerofy a pointer to the qpair in the context structure when the qpair is being deleted, so the callback
function may understand that the pointer to the qpair is no longer valid.
The above doesn't fully resolve the issue, the next patch will intoruduce a mutex per RDMA device to lock
qpair destruction process and handling of ibv_async events

Fixes #1075

Change-Id: I3a27cd50957d84b72698df73bc51f9b39debd5a8

Signed-off-by: Alexey Marchuk <alexeymar@mellanox.com>
Signed-off-by: Sasha Kotchubievsky <sashakot@mellanox.com>
Signed-off-by: Evgeniy Kochetov <evgeniik@mellanox.com>
Change-Id: I92bc598239a6688994b8010464137bad7b08bd78